[Task][HSH-302]Unit testing[Under review]
