+++ .env
@@ -3,3 +3,8 @@ MNEMONIC= outside urban recycle float ocean merge glove expose price dizzy uncov
 API_URL = https://eth-ropsten.alchemyapi.io/v2/xxxx
 HARMONY_PRIVATE_KEY= 0x1d17d2b799083e2ca539f40bdf6d3a13227c3b56a42abd787df096871c0ab153
 PRIVATE_KEY= 0x5eeacaa64c632c8291341941519ae640b316d3061702671de1b7ae6312fb0d47
+
+minUSDT = 10000000000 
+minUSDC = 10000000000
+minBTC = 10000000
+minBNB = 25000000
