+++ contracts/facets/Deposit.sol
@@ -132,7 +132,10 @@ contract Deposit is Pausable, IDeposit{
 				savingsAccount.deposits[deposit.id -1].isTimelockActivated = true;
 				savingsAccount.deposits[deposit.id -1].activationTime = block.timestamp;
 				savingsAccount.deposits[deposit.id -1].lastUpdate = block.timestamp;
+				return false;
 			}
+
+			console.log("Block time: ", block.timestamp);
 			require(deposit.activationTime + deposit.timelockValidity <= block.timestamp, "ERROR: Active timelock");
 		} 
 		ds.token = IBEP20(LibOpen._connectMarket(_market));
@@ -291,6 +294,7 @@ contract Deposit is Pausable, IDeposit{
 		LibOpen._isMarketSupported(_market);
 		ds.token = IBEP20(LibOpen._connectMarket(_market));
 		// _quantifyAmount(_market, _amount);
+		console.log("Token Address: ", address(ds.token)); 
 		LibOpen._minAmountCheck(_market, _amount);
 	}
